requirements.txt
streamlit
pandas
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
tqdm
PyMuPDF
pytesseract
pillow
openpyxl
openai
fastapi
uvicorn
itsdangerous
requests
